Revert "Change expiresIn's type from int to long"
This reverts commit 56ac955aba7e96b4ced39ef3f9ab95066de50710.

+++ trip-kit-booking/src/main/java/com/skedgo/android/tripkit/booking/AccessTokenResponse.java
@@ -13,7 +13,7 @@ public interface AccessTokenResponse {
 
   @SerializedName("access_token") String accessToken();
   @SerializedName("token_type") String tokenType();
-  @SerializedName("expires_in") long expiresIn();
+  @SerializedName("expires_in") int expiresIn();
   @Nullable @SerializedName("refresh_token") String refreshToken();
 
 }
